requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.10.8
python-telegram-bot==11.1.0
pytz==2021.3
six==1.16.0
tornado==6.1
tzlocal==3.0
requests==2.27.1